Update jose, rework auth slice and make vatsim token stored in local storage
