+++ package.json
@@ -32,7 +32,7 @@
     "eslint-plugin-react": "^7.31.11",
     "eslintrc": "^1.0.6",
     "geomag": "^1.0.0",
-    "jose": "^4.11.1",
+    "jose": "^4.15.9",
     "lodash": "^4.17.21",
     "npm": "^8.19.3",
     "react": "^18.2.0",
