+++ src/login/Login.tsx
@@ -14,6 +14,7 @@ function redirectLogin() {
 }
 
 const Login = () => {
+  console.log("initalised login")
   const dispatch = useRootDispatch();
   const navigate = useNavigate();
   const [searchParams] = useSearchParams();
